requirements.txt
fastapi
uvicorn
PyMuPDF
Pillow
python-docx
openai
python-dotenv
